# core/core-api/src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: soo-auth-service API
  description: 인증 서비스 API 가이드 (REST API Docs).
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost
...
paths:
...
  /api/v1/auth/email/signup:
    post:
      tags:
      - Auth
      summary: 이메일 회원가입
      description: "이메일/비밀번호 기반 로컬 계정을 생성합니다. 휴대폰 인증은 선택입니다.\n\n### 인증\n- 인증 불필요\n\
        \n### 요청 바디 필드\n- `email`: 이메일\n- `password`: 비밀번호\n- `name`: 이름\n- `nickname`:\
        \ 닉네임\n- `gender`: 성별 (MALE/FEMALE/UNKNOWN)\n- `phoneNumber`: 휴대폰 번호 (선택)\n\
        - `phoneVerificationToken`: 휴대폰 인증 완료 토큰 (선택)\n- `locale`: 로케일\n- `profileImageUrl`:\
        \ 프로필 이미지 URL\n- `thumbnailImageUrl`: 썸네일 이미지 URL\n- `birthyear`: 출생연도 (yyyy)\n\
        - `birthday`: 생일 (MM-DD)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한\
        \ 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger\
        \ UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-email-signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-email-signup-989332341'
            examples:
              auth-email-signup:
                value: "{\n  \"email\" : \"user@example.com\",\n  \"password\" : \"\
                  P@ssw0rd!\",\n  \"name\" : \"홍길동\",\n  \"nickname\" : \"gildong\"\
                  ,\n  \"gender\" : \"MALE\",\n  \"phoneNumber\" : null,\n  \"phoneVerificationToken\"\
                  \ : null,\n  \"locale\" : \"ko-KR\",\n  \"profileImageUrl\" : \"\
                  http://dev.soo.io/profile.png\",\n  \"thumbnailImageUrl\" : \"http://dev.soo.io/thumbnail.png\"\
                  ,\n  \"birthyear\" : \"1990\",\n  \"birthday\" : \"01-31\"\n}"
      responses:
        "200":
          description: "200"
  /api/v1/auth/id/login:
    post:
      tags:
      - Auth
      summary: 아이디 로그인
      description: "아이디/비밀번호로 로그인하고 액세스/리프레시 토큰을 발급합니다.\n\n### 인증\n- 인증 불필요\n\n###\
        \ 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### 요청 바디 필드\n- `loginId`: 로그인 아이디\n- `password`:\
        \ 비밀번호\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 토큰 정보\n- `data.accessToken`: 액세스 토큰\n- `data.accessExpiresInSec`:\
        \ 액세스 토큰 만료(초)\n- `data.refreshToken`: 리프레시 토큰\n- `data.refreshExpiresInSec`:\
        \ 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요\
        한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의\
        \ **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-id-login
      parameters:
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-id-login-1427235532'
            examples:
              auth-id-login:
                value: "{\n  \"loginId\" : \"test-user\",\n  \"password\" : \"P@ssw0rd!\"\
                  \n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-phone-login-1150988368'
              examples:
                auth-id-login:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.476006794\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/id/login\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"accessToken\"\
                    \ : \"access-token\",\n    \"accessExpiresInSec\" : 3600,\n  \
                    \  \"refreshToken\" : \"refresh-token\",\n    \"refreshExpiresInSec\"\
                    \ : 1209600\n  }\n}"
  /api/v1/auth/id/signup:
    post:
      tags:
      - Auth
      summary: 아이디 회원가입
      description: "아이디/비밀번호 계정을 생성합니다. 휴대폰 인증이 필수입니다.\n\n### 인증\n- 인증 불필요\n\n###\
        \ 요청 바디 필드\n- `loginId`: 로그인 아이디\n- `password`: 비밀번호\n- `phoneNumber`: 휴대폰\
        \ 번호\n- `phoneVerificationToken`: 휴대폰 인증 완료 토큰\n\n### 공통 에러\n- **401** 인증\
        \ 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식/필수\
        \ 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-id-signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-id-signup936474755'
            examples:
              auth-id-signup:
                value: "{\n  \"loginId\" : \"test-user\",\n  \"password\" : \"P@ssw0rd!\"\
                  ,\n  \"phoneNumber\" : \"+82 10-1234-5678\",\n  \"phoneVerificationToken\"\
                  \ : \"verified-phone-token\"\n}"
      responses:
        "200":
          description: "200"
...
  /api/v1/auth/phone/login:
    post:
      tags:
      - Auth
      summary: 휴대폰 로그인
      description: "휴대폰 인증 토큰으로 로그인하고 액세스/리프레시 토큰을 발급합니다.\n\n### 인증\n- 인증 불필요\n\n\
        ### 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### 요청 바디 필드\n- `phoneNumber`: 휴대폰 번호\
        \n- `phoneVerificationToken`: 휴대폰 인증 완료 토큰\n\n### 응답 바디 필드\n- `result`: 요청\
        \ 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`:\
        \ 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`:\
        \ 요청 경로\n- `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열\
        \ (없으면 null)\n- `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로\
        케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 토큰 정보\n- `data.accessToken`:\
        \ 액세스 토큰\n- `data.accessExpiresInSec`: 액세스 토큰 만료(초)\n- `data.refreshToken`:\
        \ 리프레시 토큰\n- `data.refreshExpiresInSec`: 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401**\
        \ 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식\
        /필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다\
        ."
      operationId: auth-phone-login
      parameters:
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-phone-login1536767407'
            examples:
              auth-phone-login:
                value: "{\n  \"phoneNumber\" : \"+821012345678\",\n  \"phoneVerificationToken\"\
                  \ : \"verified-phone-token\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-phone-login-1150988368'
              examples:
                auth-phone-login:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.351754666\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/phone/login\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"accessToken\"\
                    \ : \"access-token\",\n    \"accessExpiresInSec\" : 3600,\n  \
                    \  \"refreshToken\" : \"refresh-token\",\n    \"refreshExpiresInSec\"\
                    \ : 1209600\n  }\n}"
  /api/v1/auth/phone/signup:
    post:
      tags:
      - Auth
      summary: 휴대폰 회원가입
      description: "휴대폰 인증 토큰으로 계정을 생성합니다.\n\n### 인증\n- 인증 불필요\n\n### 요청 바디 필드\n-\
        \ `phoneNumber`: 휴대폰 번호\n- `phoneVerificationToken`: 휴대폰 인증 완료 토큰\n\n### 공\
        통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation\
        \ 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과\
        로 자동 포함됩니다."
      operationId: auth-phone-signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-phone-login1536767407'
            examples:
              auth-phone-signup:
                value: "{\n  \"phoneNumber\" : \"+82 10-1234-5678\",\n  \"phoneVerificationToken\"\
                  \ : \"verified-phone-token\"\n}"
      responses:
        "200":
          description: "200"
  /api/v1/auth/token/refresh:
    post:
      tags:
      - Auth
      summary: 토큰 재발급
      description: "리프레시 토큰으로 새 액세스/리프레시 토큰을 발급합니다.\n\n### 인증\n- 인증 불필요(리프레시 토큰 필요\
        )\n\n### 헤더\n- `X-Device-Id`: 디바이스 식별자\n\n### 요청 바디 필드\n- `refreshToken`:\
        \ 리프레시 토큰\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`:\
        \ 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`:\
        \ 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n- `meta.request.method`: HTTP 메\
        서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n- `meta.paging`: 페이징 메타 정보\
        \ (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에\
        서만 포함)\n- `data`: 토큰 정보\n- `data.accessToken`: 액세스 토큰\n- `data.accessExpiresInSec`:\
        \ 액세스 토큰 만료(초)\n- `data.refreshToken`: 리프레시 토큰\n- `data.refreshExpiresInSec`:\
        \ 리프레시 토큰 만료(초)\n\n### 공통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요\
        한 API)\n- **400** validation 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의\
        \ **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다."
      operationId: auth-token-refresh
      parameters:
      - name: X-Device-Id
        in: header
        description: 디바이스 식별자
        required: true
        schema:
          type: string
        example: device-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-token-refresh1056639717'
            examples:
              auth-token-refresh:
                value: "{\n  \"refreshToken\" : \"refresh-token\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-phone-login-1150988368'
              examples:
                auth-token-refresh:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.448222228\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/token/refresh\",\n      \"method\"\
                    \ : \"POST\"\n    }\n  },\n  \"data\" : {\n    \"accessToken\"\
                    \ : \"access-token\",\n    \"accessExpiresInSec\" : 3600,\n  \
                    \  \"refreshToken\" : \"refresh-token\",\n    \"refreshExpiresInSec\"\
                    \ : 1209600\n  }\n}"
...
components:
  schemas:
...
    api-v1-auth-email-signup-989332341:
      required:
      - email
      - gender
      - password
      type: object
      properties:
        birthday:
          type: string
          description: 생일 (MM-DD)
          nullable: true
        phoneVerificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰 (선택)
          nullable: true
        thumbnailImageUrl:
          type: string
          description: 썸네일 이미지 URL
          nullable: true
        password:
          type: string
          description: 비밀번호
        phoneNumber:
          type: string
          description: 휴대폰 번호 (선택)
          nullable: true
        birthyear:
          type: string
          description: 출생연도 (yyyy)
          nullable: true
        gender:
          type: string
          description: 성별 (MALE/FEMALE/UNKNOWN)
        nickname:
          type: string
          description: 닉네임
          nullable: true
        name:
          type: string
          description: 이름
          nullable: true
        locale:
          type: string
          description: 로케일
          nullable: true
        profileImageUrl:
          type: string
          description: 프로필 이미지 URL
          nullable: true
        email:
          type: string
          description: 이메일
...
    api-v1-auth-id-signup936474755:
      required:
      - loginId
      - password
      - phoneNumber
      - phoneVerificationToken
      type: object
      properties:
        phoneVerificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
        password:
          type: string
          description: 비밀번호
        phoneNumber:
          type: string
          description: 휴대폰 번호
        loginId:
          type: string
          description: 로그인 아이디
...
    api-v1-auth-phone-login-1150988368:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - accessExpiresInSec
          - accessToken
          - refreshExpiresInSec
          - refreshToken
          type: object
          properties:
            refreshExpiresInSec:
              type: number
              description: 리프레시 토큰 만료(초)
            accessExpiresInSec:
              type: number
              description: 액세스 토큰 만료(초)
            accessToken:
              type: string
              description: 액세스 토큰
            refreshToken:
              type: string
              description: 리프레시 토큰
          description: 토큰 정보
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...
    api-v1-auth-id-login-1427235532:
      required:
      - loginId
      - password
      type: object
      properties:
        password:
          type: string
          description: 비밀번호
        loginId:
          type: string
          description: 로그인 아이디
...
    api-v1-auth-local-token-refresh1056639717:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
          description: 리프레시 토큰
...
    api-v1-auth-phone-login1536767407:
      required:
      - phoneNumber
      - phoneVerificationToken
      type: object
      properties:
        phoneVerificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
        phoneNumber:
          type: string
          description: 휴대폰 번호
...